+++ src/handle_evid.cpp
@@ -32,24 +32,49 @@ extern "C" void handleTlast(double *time, rx_solving_options_ind *ind) {
 }
 
 // Linear compartment models/functions
+// Note: Rf_errorcall is not thread-safe and cannot be called from
+// within OpenMP parallel regions.  In parallel regions, return NA_REAL and
+// set badSolve so the error is handled after the parallel region completes.
+// In single-threaded context, call Rf_errorcall for an actionable error message.
 extern "C" double _getDur(int l, rx_solving_options_ind *ind, int backward, unsigned int *p) {
   double dose = getDoseNumber(ind, l);
   if (backward==1 && l != 0){
     if (l <= 0) {
-      (Rf_errorcall)(R_NilValue, _("could not find a start to the infusion #1"));
+      rx_solving_options *op = &op_global;
+      if (omp_in_parallel()) {
+        int newBadSolve = 1;
+#pragma omp atomic write
+        op->badSolve = newBadSolve;
+        return NA_REAL;
+      }
+      (Rf_errorcall)(R_NilValue, "infusion start cannot be found (l <= 0)");
     }
     p[0] = l-1;
     while (p[0] > 0 && getDoseNumber(ind, p[0]) != -dose){
       p[0]--;
     }
     if (getDoseNumber(ind, p[0]) != -dose){
-      (Rf_errorcall)(R_NilValue, _("could not find a start to the infusion #2"));
+      rx_solving_options *op = &op_global;
+      if (omp_in_parallel()) {
+        int newBadSolve = 1;
+#pragma omp atomic write
+        op->badSolve = newBadSolve;
+        return NA_REAL;
+      }
+      (Rf_errorcall)(R_NilValue, "infusion start cannot be found");
     }
     return getAllTimes(ind, ind->idose[l]) - getAllTimes(ind, ind->idose[p[0]]);
   } else {
     if (l >= ind->ndoses) {
       if (backward==2) return(NA_REAL);
-      (Rf_errorcall)(R_NilValue, _("could not find an end to the infusion"));
+      rx_solving_options *op = &op_global;
+      if (omp_in_parallel()) {
+        int newBadSolve = 1;
+#pragma omp atomic write
+        op->badSolve = newBadSolve;
+        return NA_REAL;
+      }
+      (Rf_errorcall)(R_NilValue, "infusion end cannot be found (l >= ndoses)");
     }
     p[0] = l+1;
     while (p[0] < ind->ndoses && getDoseNumber(ind, p[0]) != -dose){
@@ -57,7 +82,14 @@ extern "C" double _getDur(int l, rx_solving_options_ind *ind, int backward, unsi
     }
     if (getDoseNumber(ind, p[0]) != -dose){
       if (backward==2) return(NA_REAL);
-      (Rf_errorcall)(R_NilValue, _("could not find an end to the infusion"));
+      rx_solving_options *op = &op_global;
+      if (omp_in_parallel()) {
+        int newBadSolve = 1;
+#pragma omp atomic write
+        op->badSolve = newBadSolve;
+        return NA_REAL;
+      }
+      (Rf_errorcall)(R_NilValue, "infusion end cannot be found");
     }
     return getAllTimes(ind, ind->idose[p[0]]) - getAllTimes(ind, ind->idose[l]);
   }